+++ Tests/BlueprintTests/Session.swift
@@ -17,10 +17,61 @@ final class SessionSpec: QuickSpec {
 
                 var currentId: TestId? = nil
                 session.user
-                    .subscribe(onNext: { currentId = $0 })
+                    .subscribe(onNext: { user in
+                        switch user {
+                        case let .success(id): currentId = id
+                        case .failure: currentId = nil
+                        }
+                    })
                     .disposed(by: disposeBag)
-                session.update(id)
+                session.update(.success(id))
                 expect(currentId).toEventually(equal(id))
+            }
+
+            context("on receiving error") {
+                it("should publish nil id to subscribers") {
+                    let disposeBag = DisposeBag()
+                    let session = InMemorySession<TestId>()
+                    let id = TestId("test")
+
+                    var currentId: TestId? = nil
+                    session.user
+                        .subscribe(onNext: { user in
+                            switch user {
+                            case let .success(id): currentId = id
+                            case .failure: currentId = nil
+                            }
+                        })
+                        .disposed(by: disposeBag)
+
+                    session.update(.success(id))
+                    expect(currentId).toEventually(equal(id))
+
+                    session.update(.failure(TestError()))
+                    expect(currentId).toEventually(beNil())
+                }
+            }
+
+            context("after receiving errors") {
+                it("can update the current user") {
+                    let disposeBag = DisposeBag()
+                    let session = InMemorySession<TestId>()
+                    let id = TestId("test")
+
+                    var currentId: TestId? = nil
+                    session.user
+                        .subscribe(onNext: { user in
+                            switch user {
+                            case let .success(id): currentId = id
+                            case .failure: currentId = nil
+                            }
+                        })
+                        .disposed(by: disposeBag)
+
+                    session.update(.failure(TestError()))
+                    session.update(.success(id))
+                    expect(currentId).toEventually(equal(id))
+                }
             }
         }
     }
@@ -31,5 +82,8 @@ final class SessionSpec: QuickSpec {
         init(_ value: String) {
             self.value = value
         }
+    }
+
+    struct TestError: Error {
     }
 }